+++ packages/web/src/common/components/BannerCarousel/types.ts
@@ -1,0 +1,5 @@
+import { CampaignCardProps } from '../CampaignCard/types';
+
+export type BannerCarouselPropers = {
+  banners: CampaignCardProps[];
+};