+++ packages/web/src/common/components/SwiperCarousel/types.ts
@@ -2,17 +2,12 @@ export type CarouselProps = {
   children: React.ReactNode[] | React.ReactNode;
 
   // Display
-  hideArrows?: boolean;
-  hidePillButtons?: boolean;
-  loop?: boolean;
+  pagination?: boolean;
+  navigation?: boolean;
 
   // Breakpoints
   elementsPerPageDesktop?: number;
   elementsPerPageLaptop?: number;
   elementsPerPageTablet?: number;
   elementsPerPageMobile?: number;
-
-  // Auto play
-  autoPlay?: boolean;
-  autoPlayIntervalMs?: number;
 };